+++ apps/mmi/src/App.test.tsx
@@ -1,0 +1,5 @@
+import App from './App'
+
+test('render App', () => {
+  // render(<App />)
+})