+++ apps/safe-claiming-app/src/hooks/useFetchVestingStatus.tsx
@@ -1,58 +1,53 @@
 import { useSafeAppsSDK } from "@gnosis.pm/safe-apps-react-sdk"
-import { useEffect, useMemo, useState } from "react"
 import { ZERO_ADDRESS } from "src/config/constants"
+import useAsync, { type AsyncResult } from "src/hooks/useAsync"
 import { Airdrop__factory } from "src/types/contracts/factories/Airdrop__factory"
 import { VestingStatus } from "src/types/vestings"
 import { getWeb3Provider } from "src/utils/getWeb3Provider"
+import { ethers } from "ethers"
+import { useMemo } from "react"
+
+const getVestingStatus = async (
+  airdropAddress: string | undefined,
+  web3Provider: ethers.providers.Provider,
+  vestingId: string | undefined
+): Promise<VestingStatus | undefined> => {
+  if (!vestingId || !airdropAddress) {
+    return
+  }
+
+  try {
+    // Check if vesting is already redeemed and amount claimed
+    const newVesting = await Airdrop__factory.connect(
+      airdropAddress,
+      web3Provider
+    ).vestings(vestingId)
+
+    if (newVesting.account.toLowerCase() === ZERO_ADDRESS.toLowerCase()) {
+      return {
+        isRedeemed: false,
+        amountClaimed: "0",
+      }
+    }
+    return {
+      amountClaimed: newVesting.amountClaimed.toString(),
+      isRedeemed: true,
+    }
+  } catch (error) {
+    console.error(error)
+    throw Error("Error fetching vesting status")
+  }
+}
 
 export const useFetchVestingStatus = (
   vestingId: string | undefined,
-  airdropAddress: string
-): [VestingStatus | undefined, string | undefined] => {
+  airdropAddress: string | undefined
+): AsyncResult<VestingStatus> => {
   const { safe, sdk } = useSafeAppsSDK()
   const web3Provider = useMemo(() => getWeb3Provider(safe, sdk), [safe, sdk])
-  const [vestingStatus, setVestingStatus] = useState<VestingStatus>()
-  const [error, setError] = useState<string>()
 
-  useEffect(() => {
-    let isMounted = true
-    const fetchVestingStatus = async () => {
-      if (!vestingId) {
-        return
-      }
-      try {
-        // Check if vesting is already redeemed and amount claimed
-        const newVesting = await Airdrop__factory.connect(
-          airdropAddress,
-          web3Provider
-        ).vestings(vestingId)
-
-        if (newVesting.account.toLowerCase() === ZERO_ADDRESS.toLowerCase()) {
-          isMounted &&
-            setVestingStatus({
-              isRedeemed: false,
-              amountClaimed: "0",
-            })
-        } else {
-          isMounted &&
-            setVestingStatus({
-              amountClaimed: newVesting.amountClaimed.toString(),
-              isRedeemed: true,
-            })
-        }
-        setError(undefined)
-      } catch (error) {
-        console.error(error)
-        setError("Fetching vesting status failed.")
-      }
-    }
-
-    fetchVestingStatus()
-
-    return () => {
-      isMounted = false
-    }
-  }, [airdropAddress, vestingId, web3Provider])
-
-  return [vestingStatus, error]
+  return useAsync(
+    () => getVestingStatus(airdropAddress, web3Provider, vestingId),
+    [airdropAddress, vestingId, web3Provider]
+  )
 }