+++ apps/wallet-connect/src/hooks/useWalletConnectV1.tsx
@@ -5,6 +5,9 @@ import { IClientMeta, IWalletConnectSession } from '@walletconnect/legacy-types'
 import { useSafeAppsSDK } from '@gnosis.pm/safe-apps-react-sdk'
 import { SafeAppProvider } from '@gnosis.pm/safe-apps-provider'
 
+import { useApps } from './useApps'
+import { trackSafeAppEvent } from '../utils/analytics'
+
 const rejectWithMessage = (connector: WalletConnect, id: number | undefined, message: string) => {
   connector.rejectRequest({ id, error: { message } })
 }
@@ -18,7 +21,20 @@ const useWalletConnectV1 = () => {
     [sdk, safe],
   )
 
+  const { findSafeApp } = useApps()
+
   const localStorageSessionKey = useRef(`session_${safe.safeAddress}`)
+
+  const trackEvent = useCallback(
+    (action, meta) => {
+      if (!meta) return
+
+      const safeApp = meta && findSafeApp(meta.url)
+
+      trackSafeAppEvent(action, safeApp?.name || meta?.url)
+    },
+    [findSafeApp],
+  )
 
   const wcDisconnect = useCallback(async () => {
     try {
@@ -50,6 +66,8 @@ const useWalletConnectV1 = () => {
           chainId: safe.chainId,
         })
 
+        trackEvent('New session', wcConnector.peerMeta)
+
         setWcClientData(payload.params[0].peerMeta)
       })
 
@@ -60,6 +78,8 @@ const useWalletConnectV1 = () => {
 
         try {
           let result = await web3Provider.send(payload.method, payload.params)
+
+          trackEvent('Transaction Confirmed', wcConnector.peerMeta)
 
           wcConnector.approveRequest({
             id: payload.id,
@@ -77,7 +97,7 @@ const useWalletConnectV1 = () => {
         wcDisconnect()
       })
     },
-    [safe, wcDisconnect, web3Provider],
+    [safe, wcDisconnect, web3Provider, trackEvent],
   )
 
   useEffect(() => {